+++ app/schema/passageiro.py
@@ -1,21 +1,20 @@
 from typing import Optional
 from pydantic import BaseModel
-from sqlalchemy import BigInteger, Float
 
 class PassageiroSchema(BaseModel):
-    id_passageiro: BigInteger
-    id_usuario: BigInteger
+    id_passageiro: int
+    id_usuario: int
 
-    media_avaliacao: Optional[Float]
+    media_avaliacao: Optional[float]
 
     class Config:
         from_attributes = True
 
 class PassageiroUpdateSchema(BaseModel):
-    id_passageiro: Optional[BigInteger]
-    id_usuario: Optional[BigInteger]
+    id_passageiro: Optional[int]
+    id_usuario: Optional[int]
 
-    media_avaliacao: Optional[Float]
+    media_avaliacao: Optional[float]
 
     class Config:
         from_attributes = True
